feat(portal): inline RepairBuilderSection shared section
